--- src/test/java/se/iths/corkdork/CorkDorkApplicationTests.java
@@ -1,13 +1,0 @@
-package se.iths.corkdork;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest(classes = CorkDorkApplicationTests.class)
-class CorkDorkApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
