+++ src/test/java/se/iths/corkdork/modellmapper/ModelMapperTest.java
@@ -1,4 +1,4 @@
-package se.iths.corkdork;
+package se.iths.corkdork.modellmapper;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
